Add files via upload
This file contains a prototype Jinja2 setup that uses a python script that creates a txt file using a jinja2 template and a csv file that contains specific information. This prototype exists to prove we can take a template of a configuration device, modify it, and create a new configuration file with the new modifications.